Add files via upload
An application which allons to share contents among users. When they are registered, they can be Owner who onlined contents and can sell it; or Customer who need contents. 
Customers can look at all the available contents on the platform.

This program just implements the Owner and His content registration. This content will be a book.
Then, the Customers will be registered too, and he could know about the content available.